seperate map and core
